risteys: fix median being an int (+ db migration)
Median can be a float, for example: Median([1, 2]) = 1.5.

GitHub: fixes #37

+++ risteys_elixir/lib/risteys/stats_sex.ex
@@ -8,7 +8,7 @@ defmodule Risteys.StatsSex do
 
     field :case_fatality, :float
     field :mean_age, :float
-    field :median_reoccurence, :integer
+    field :median_reoccurence, :float
     field :n_individuals, :integer
     field :prevalence, :float
     field :reoccurence_rate, :float
